Database.HsSqlPpp.Annotation

Description

Contains the annotation data types and a few auxiliary functions.

Annotation data types

data Annotation

Annotation type - one of these is attached to most of the data types used in the ast.

Constructors

Annotation

Fields

asrc :: Maybe SourcePosition

atype :: Maybe Type

errs :: [TypeError]

stType :: Maybe StatementType

catUpd :: [CatalogUpdate]

fnProt :: Maybe FunctionPrototype

infType :: Maybe Type

type SourcePosition = (String, Int, Int)

type StatementType = ([Type], [(String, Type)])

getAnnotation :: Data a => a -> Annotation

get the annotation for the root element of the tree passed

updateAnnotations :: Data a => (Annotation -> Annotation) -> a -> a

updateAnnotation :: Data a => (Annotation -> Annotation) -> a -> a

Update all the annotations in a tree using the function supplied

getAnnotations :: Data a => a -> [Annotation]

emptyAnnotation :: Annotation

resetAnnotations :: Data a => a -> a