Database.HsSqlPpp.Catalog

Description

This module contains the database catalog data types and helper functions.

The catalog data type serves the following purposes:

Contains all the catalog information needed to type check against an existing database.

A copy of the catalog information from a default template1 database is included - defaultTemplate1Catalog.

It is used internally to keep track of updates to the catalog whilst running an annotation process (e.g. so that a select can type check against a create table given in the same source). It is also used to track other identifier types, such as attribute references in select expressions, and argument and variable types inside create function statements.

You can see what kind of stuff is contained in the Catalog type by looking at the CatalogUpdate type.

Data types

data Catalog

The main datatype, this holds the catalog and context information to type check against.

Updates

data CatalogUpdate

Constructors

CatCreateScalar Type String Bool
add a new scalar type with the name given, also creates an array type automatically

CatCreateDomain Type Type

CatCreateComposite String [(String, Type)]

CatCreateCast Type Type CastContext

CatCreateTable String [(String, Type)] [(String, Type)]

CatCreateView String [(String, Type)]

CatCreateFunction FunFlav String [Type] Type Bool

CatDropFunction Bool String [Type]

ppCatUpdate :: CatalogUpdate -> String

bits and pieces

data CastContext

Use to note what the flavour of a cast is, i.e. if/when it can be used implicitly.

Constructors

ImplicitCastContext

AssignmentCastContext

ExplicitCastContext

data CompositeFlavour

Used to distinguish between standalone composite types, and automatically generated ones, generated from a table or view respectively.

Constructors

Composite

TableComposite

ViewComposite

type CompositeDef = (String, CompositeFlavour, Type, Type)

Provides the definition of a composite type. The components are composite (or table or view) name, the flavour of the composite, the types of the composite attributes, and the types of the system columns iff the composite represents a table type (the third and fourth components are always CompositeTypes).

type FunctionPrototype = (String, [Type], Type, Bool)

The components are: function (or operator) name, argument types, return type and is variadic.

type DomainDefinition = (Type, Type)

The components are domain type, base type (todo: add check constraint).

data FunFlav

Constructors

FunPrefix

FunPostfix

FunBinary

FunName

FunAgg

FunWindow

Catalog values

emptyCatalog :: Catalog

Represents an empty catalog. This doesn't contain things like the 'and' operator, defaultCatalog contains these.

defaultCatalog :: Catalog

Represents what you probably want to use as a starting point if you are building an catalog from scratch. It contains information on built in function like things that aren't in the PostgreSQL catalog, such as greatest, coalesce, keyword operators like 'and', etc..

defaultTemplate1Catalog :: Catalog

Catalog comparison

data CatalogDiff

items in first catalog and not second, items in second and not first.

Constructors

CatalogDiff [CatalogUpdate] [CatalogUpdate]

compareCatalogs :: Catalog -> Catalog -> Catalog -> CatalogDiff

find differences between two catalogs

ppCatDiff :: CatalogDiff -> String

print a catdiff in a more human readable way than show.

Functions

updateCatalog :: Catalog -> [CatalogUpdate] -> Either [TypeError] Catalog

Applies a list of CatalogUpdates to an Catalog value to produce a new Catalog value.

deconstructCatalog :: Catalog -> [CatalogUpdate]

operator utils

data OperatorType

Constructors

BinaryOp

PrefixOp

PostfixOp

getOperatorType :: Catalog -> String -> Either [TypeError] OperatorType

isOperatorName :: String -> Bool