Database.HsSqlPpp.Parse

Description

Functions to parse SQL.

Main

parseStatements

Arguments

:: ParseFlags
parse options

-> FilePath
filename to use in errors

-> Maybe (Int, Int)
set the line number and column number of the first char in the source (used in annotation)

-> Text
a string containing the sql to parse

-> Either ParseErrorExtra [Statement]

Parse a list of statements

parseProcSQL

Arguments

:: ParseFlags
parse options

-> FilePath
filename to use in errors

-> Maybe (Int, Int)
set the line number and column number

-> Text
a string containing the sql to parse

-> Either ParseErrorExtra [Statement]

Parse a list of procedural sql statements

parseQueryExpr

Arguments

:: ParseFlags
parse options

-> FilePath
filename to use in errors

-> Maybe (Int, Int)
set the line number and column number

-> Text
a string containing the sql to parse

-> Either ParseErrorExtra QueryExpr

Parse a single query expr

parseScalarExpr

Arguments

:: ParseFlags
parse options

-> FilePath
filename to use in errors

-> Maybe (Int, Int)
set the line number and column number

-> Text
a string containing the sql to parse

-> Either ParseErrorExtra ScalarExpr

Parse a single scalar expr

Parsing options

data ParseFlags

Settings to influence the parsing

Constructors

ParseFlags

Fields

pfDialect :: Dialect

defaultParseFlags :: ParseFlags

data Dialect

Constructors

Dialect

Fields

diName :: String

diSyntaxFlavour :: SyntaxFlavour

diCanonicalTypeNames :: [(Text, [Text])]

diTextTypes :: [Text]

diDatetimeTypes :: [Text]

diNumberTypes :: [Text]

namesForAnsiTypes :: [(Text, Text)]

diDefaultCatalog :: Catalog

ansiDialect :: Dialect

errors

data ParseErrorExtra

Simple wrapper to allow showing the source context of a ParseError

Constructors

ParseErrorExtra

Fields

parseErrorError :: ParseError

wrapped error

parseErrorPosition :: Maybe (Int, Int)

source position adjustment to get the context bit in error messages right - this is the same as what is passed into parseSqlWithPosition

parseErrorSqlSource :: Text

sql source

internals

parseName :: ParseFlags -> FilePath -> Maybe (Int, Int) -> Text -> Either ParseErrorExtra Name

parseNameComponent :: ParseFlags -> FilePath -> Maybe (Int, Int) -> Text -> Either ParseErrorExtra NameComponent