Database.HsSqlPpp.SqlQuote

Description

A quasiquoter for SQL. Antiquoting is a bit inconsistent. The splice variable names must be all lower case because of a limitation in the parser.

Example:

 
 import Database.HsSqlPpp.Ast
 import Database.HsSqlPpp.SqlQuote
 import Database.HsSqlPpp.Annotation

 test :: Statement
 test = [$sqlStmt|

   create table $(tablename) (
    $(varname) $(typename)
   );

         |]
   where
     tablename = "my_table"
     varname = "my_field"
     typename = "text"

See http://jakewheat.github.com/hssqlppp/QuasiQuoteTests.html for more simple examples, and http://jakewheat.github.com/hssqlppp/source/examples/Database/HsSqlPpp/Examples/Extensions/ for some example files which use quasiquotation to do ast transformations which implement syntax extensions to sql

Documentation

sqlStmts :: QuasiQuoter

parses Statements

sqlStmt :: QuasiQuoter

parses a single Statement

pgsqlStmts :: QuasiQuoter

parses plpgsql Statements

pgsqlStmt :: QuasiQuoter

parses a plpgsql Statement

sqlExpr :: QuasiQuoter

parse an Expression