Database.HsSqlPpp.TypeCheck

Description

Contains functions for typechecking sql asts.

typechecking/ annotation functions

typeCheckStatements :: TypeCheckFlags -> Catalog -> [Statement] -> (Catalog, [Statement])

Typechecks the ast, and returns the updated catalog (which includes changes from any ddl statements in the ast).

typeCheckQueryExpr :: TypeCheckFlags -> Catalog -> QueryExpr -> QueryExpr

Typecheck a query expr

typeCheckScalarExpr :: TypeCheckFlags -> Catalog -> Environment -> ScalarExpr -> ScalarExpr

Not working yet. Typechecks a statement possibly containing ? placeholders. These are annotated with the 'inferred type', and the stType annotation on the return value can be used to get this info easily. Returns Left if the statement is not a query,insert,update or delete statement

data TypeCheckFlags

some options when typechecking

Constructors

TypeCheckFlags

Fields

tcfAddQualifiers :: Bool

add qualifiers to unqualified ids where possible

tcfAddFullTablerefAliases :: Bool

add full aliases to every tableref and subselect

tcfAddSelectItemAliases :: Bool

add explicit aliases to all select items

tcfExpandStars :: Bool

expand stars in select lists to explicit list columns

tcfDialect :: Dialect

data Dialect

Constructors

Dialect

Fields

diName :: String

diSyntaxFlavour :: SyntaxFlavour

diCanonicalTypeNames :: [(Text, [Text])]

diTextTypes :: [Text]

diDatetimeTypes :: [Text]

diNumberTypes :: [Text]

namesForAnsiTypes :: [(Text, Text)]

diDefaultCatalog :: Catalog

ansiDialect :: Dialect

defaultTypeCheckFlags :: TypeCheckFlags

reasonable defaults for type checking, doesn't add anything optional

emptyEnvironment :: Environment