Database.HsSqlPpp.Ast

Description

This module contains the ast node data types. They are very permissive, in that they allow a lot of invalid SQL to be represented. The type checking process should catch all invalid trees, but doesn't quite manage at the moment. Sorry about all the seemingly pointless type synonyms below, they are an artefact of using UUAGC. You can see labels for the fields by looking at the ag source here: http://jakewheat.github.com/hssqlppp/source/src/Database/HsSqlPpp/AstInternals/AstInternal.ag.html

Main nodes

type StatementList = [Statement]

data Statement

Constructors

AlterSequence Annotation String SQIdentifier

AlterTable Annotation String AlterTableActionList

Assignment Annotation ScalarExpr ScalarExpr

Block Annotation (Maybe String) VarDefList StatementList

CaseStatement Annotation ScalarExprListStatementListPairList StatementList

CaseStatementSimple Annotation ScalarExpr ScalarExprListStatementListPairList StatementList

ContinueStatement Annotation (Maybe String)

Copy Annotation String [String] CopySource

CopyData Annotation String

CreateDomain Annotation String TypeName String MaybeBoolExpr

CreateFunction Annotation String ParamDefList TypeName Replace Language FnBody Volatility

CreateLanguage Annotation String

CreateSequence Annotation String Integer Integer Integer Integer Integer

CreateTable Annotation String AttributeDefList ConstraintList

CreateTableAs Annotation String QueryExpr

CreateTrigger Annotation String TriggerWhen [TriggerEvent] String TriggerFire String ScalarExprList

CreateType Annotation String TypeAttributeDefList

CreateView Annotation String (Maybe [String]) QueryExpr

Delete Annotation SQIdentifier TableRefList MaybeBoolExpr MaybeSelectList

DropFunction Annotation IfExists StringTypeNameListPairList Cascade

DropSomething Annotation DropType IfExists [String] Cascade

Execute Annotation ScalarExpr

ExitStatement Annotation (Maybe String)

ForIntegerStatement Annotation (Maybe String) ScalarExpr ScalarExpr ScalarExpr StatementList

ForQueryStatement Annotation (Maybe String) ScalarExpr QueryExpr StatementList

If Annotation ScalarExprStatementListPairList StatementList

Insert Annotation SQIdentifier [String] QueryExpr MaybeSelectList

Into Annotation Bool ScalarExprList Statement

LoopStatement Annotation (Maybe String) StatementList

Notify Annotation String

NullStatement Annotation

Perform Annotation ScalarExpr

QueryStatement Annotation QueryExpr

Raise Annotation RaiseType String ScalarExprList

Return Annotation MaybeScalarExpr

ReturnNext Annotation ScalarExpr

ReturnQuery Annotation QueryExpr

Set Annotation String [SetValue]

Truncate Annotation [String] RestartIdentity Cascade

Update Annotation SQIdentifier ScalarExprList TableRefList MaybeBoolExpr MaybeSelectList

WhileStatement Annotation (Maybe String) ScalarExpr StatementList

data ScalarExpr

Constructors

BooleanLit Annotation Bool

Case Annotation CaseScalarExprListScalarExprPairList MaybeScalarExpr

CaseSimple Annotation ScalarExpr CaseScalarExprListScalarExprPairList MaybeScalarExpr

Cast Annotation ScalarExpr TypeName

Exists Annotation QueryExpr

Extract Annotation ExtractField ScalarExpr

FunCall Annotation String ScalarExprList

Identifier Annotation String

InPredicate Annotation ScalarExpr Bool InList

Interval Annotation String IntervalField (Maybe Int)

LiftOperator Annotation String LiftFlavour ScalarExprList

NullLit Annotation

NumberLit Annotation String

Placeholder Annotation

PositionalArg Annotation Integer

QIdentifier Annotation ScalarExpr String

ScalarSubQuery Annotation QueryExpr

StringLit Annotation String

TypedStringLit Annotation TypeName String

WindowFn Annotation ScalarExpr ScalarExprList ScalarExprList Direction FrameClause

data QueryExpr

Constructors

CombineQueryExpr Annotation CombineType QueryExpr QueryExpr

Select Annotation Distinct SelectList TableRefList MaybeBoolExpr ScalarExprList MaybeBoolExpr ScalarExprDirectionPairList MaybeScalarExpr MaybeScalarExpr

Values Annotation ScalarExprListList

WithQueryExpr Annotation WithQueryList QueryExpr

Components

Selects

data SelectList

Constructors

SelectList Annotation SelectItemList

data SelectItem

Constructors

SelExp Annotation ScalarExpr

SelectItem Annotation ScalarExpr String

data TableRef

Constructors

FunTref Annotation ScalarExpr TableAlias

JoinTref Annotation TableRef Natural JoinType TableRef OnExpr TableAlias

SubTref Annotation QueryExpr TableAlias

Tref Annotation SQIdentifier TableAlias

data TableAlias

Constructors

FullAlias Annotation String [String]

NoAlias Annotation

TableAlias Annotation String

data JoinExpr

Constructors

JoinOn Annotation ScalarExpr

JoinUsing Annotation [String]

data JoinType

Constructors

Inner

LeftOuter

RightOuter

FullOuter

Cross

data Natural

Constructors

Natural

Unnatural

data CombineType

Constructors

Except

Union

Intersect

UnionAll

data Direction

Constructors

Asc

Desc

data Distinct

Constructors

Distinct

Dupes

data InList

Constructors

InList Annotation ScalarExprList

InQueryExpr Annotation QueryExpr

data LiftFlavour

Constructors

LiftAny

LiftAll

data FrameClause

Constructors

FrameUnboundedPreceding

FrameUnboundedFull

FrameRowsUnboundedPreceding

type WithQueryList = [WithQuery]

data WithQuery

Constructors

WithQuery Annotation String (Maybe [String]) QueryExpr

data IntervalField

Constructors

IntervalYear

IntervalMonth

IntervalDay

IntervalHour

IntervalMinute

IntervalSecond

IntervalYearToMonth

IntervalDayToHour

IntervalDayToMinute

IntervalDayToSecond

IntervalHourToMinute

IntervalHourToSecond

IntervalMinuteToSecond

data ExtractField

Constructors

ExtractCentury

ExtractDay

ExtractDecade

ExtractDow

ExtractDoy

ExtractEpoch

ExtractHour

ExtractIsodow

ExtractIsoyear

ExtractMicroseconds

ExtractMillennium

ExtractMilliseconds

ExtractMinute

ExtractMonth

ExtractQuarter

ExtractSecond

ExtractTimezone

ExtractTimezoneHour

ExtractTimezoneMinute

ExtractWeek

ExtractYear

data SQIdentifier

Constructors

SQIdentifier Annotation [String]

dml

data CopySource

Constructors

CopyFilename String

Stdin

data RestartIdentity

Constructors

RestartIdentity

ContinueIdentity

ddl

data AttributeDef

Constructors

AttributeDef Annotation String TypeName MaybeScalarExpr RowConstraintList

data RowConstraint

Constructors

NotNullConstraint Annotation String

NullConstraint Annotation String

RowCheckConstraint Annotation String ScalarExpr

RowPrimaryKeyConstraint Annotation String

RowReferenceConstraint Annotation String String (Maybe String) Cascade Cascade

RowUniqueConstraint Annotation String

data Constraint

Constructors

CheckConstraint Annotation String ScalarExpr

PrimaryKeyConstraint Annotation String [String]

ReferenceConstraint Annotation String [String] String [String] Cascade Cascade

UniqueConstraint Annotation String [String]

data AlterTableAction

Constructors

AddConstraint Annotation Constraint

AlterColumnDefault Annotation String ScalarExpr

data TypeAttributeDef

Constructors

TypeAttDef Annotation String TypeName

data TypeName

Constructors

ArrayTypeName Annotation TypeName

Prec2TypeName Annotation String Integer Integer

PrecTypeName Annotation String Integer

SetOfTypeName Annotation TypeName

SimpleTypeName Annotation String

data DropType

Constructors

Table

Domain

View

Type

data IfExists

Constructors

Require

IfExists

data Replace

Constructors

Replace

NoReplace

data Cascade

Constructors

Cascade

Restrict

data TriggerWhen

Constructors

TriggerBefore

TriggerAfter

data TriggerEvent

Constructors

TInsert

TUpdate

TDelete

data TriggerFire

Constructors

EachRow

EachStatement

functions

data FnBody

Constructors

PlpgsqlFnBody Annotation Statement

SqlFnBody Annotation StatementList

data ParamDef

Constructors

ParamDef Annotation String TypeName

ParamDefTp Annotation TypeName

data VarDef

Constructors

ParamAlias Annotation String Integer

VarAlias Annotation String String

VarDef Annotation String TypeName (Maybe ScalarExpr)

data RaiseType

Constructors

RNotice

RException

RError

data Volatility

Constructors

Volatile

Stable

Immutable

data Language

Constructors

Sql

Plpgsql

misc

data SetValue

Constructors

SetStr Annotation String

SetId Annotation String

SetNum Annotation Double

typedefs

type ScalarExprListStatementListPairList = [ScalarExprListStatementListPair]

type ScalarExprListStatementListPair = (ScalarExprList, StatementList)

type ScalarExprList = [ScalarExpr]

type MaybeSelectList = Maybe SelectList

type ParamDefList = [ParamDef]

type AttributeDefList = [AttributeDef]

type ConstraintList = [Constraint]

type TypeAttributeDefList = [TypeAttributeDef]

type TypeNameList = [TypeName]

type StringTypeNameListPair = (String, TypeNameList)

type StringTypeNameListPairList = [StringTypeNameListPair]

type ScalarExprStatementListPairList = [ScalarExprStatementListPair]

type CaseScalarExprListScalarExprPairList = [CaseScalarExprListScalarExprPair]

type MaybeScalarExpr = Maybe ScalarExpr

type MaybeBoolExpr = Maybe ScalarExpr

type TableRefList = [TableRef]

type ScalarExprListList = [ScalarExprList]

type SelectItemList = [SelectItem]

type OnExpr = Maybe JoinExpr

type RowConstraintList = [RowConstraint]

type VarDefList = [VarDef]

type ScalarExprStatementListPair = (ScalarExpr, StatementList)

type CaseScalarExprListScalarExprPair = (ScalarExprList, ScalarExpr)

type ScalarExprDirectionPair = (ScalarExpr, Direction)

type ScalarExprDirectionPairList = [ScalarExprDirectionPair]

type AlterTableActionList = [AlterTableAction]