Database.HsSqlPpp.Ast

Description

This module contains the ast node data types. They are very permissive, in that they allow a lot of invalid SQL to be represented. The type checking process should catch all invalid trees, but doesn't quite manage at the moment. Sorry about all the seemingly pointless type synonyms below, they are an artefact of using UUAGC. You can see labels for the fields by looking at the ag source here: https://github.com/JakeWheat/hssqlppp/blob/master/src/Database/HsSqlPpp/Internals/AstInternal.ag

Name and TypeName

data Name

Constructors

Name Annotation [NameComponent]

AntiName String

data NameComponent

Constructors

Nmc String

QNmc String

AntiNameComponent String

nameComponents :: Name -> [NameComponent]

ncStr :: NameComponent -> String

data TypeName

Constructors

SimpleTypeName Annotation Name

PrecTypeName Annotation Name Integer

Prec2TypeName Annotation Name Integer Integer

ArrayTypeName Annotation TypeName

SetOfTypeName Annotation TypeName

Scalar expressions

data ScalarExpr

Constructors

NumberLit Annotation String

StringLit Annotation String

NullLit Annotation

BooleanLit Annotation Bool

TypedStringLit Annotation TypeName String

Interval Annotation String IntervalField (Maybe Int)

Extract Annotation ExtractField ScalarExpr

PositionalArg Annotation Integer

Placeholder Annotation

Cast Annotation ScalarExpr TypeName

ImplicitCast Annotation ScalarExpr TypeExtra

Star Annotation

QStar Annotation NameComponent

Identifier Annotation Name

Case Annotation CaseScalarExprListScalarExprPairList MaybeScalarExpr

CaseSimple Annotation ScalarExpr CaseScalarExprListScalarExprPairList MaybeScalarExpr

Exists Annotation QueryExpr

App Annotation Name ScalarExprList

PrefixOp Annotation Name ScalarExpr

PostfixOp Annotation Name ScalarExpr

BinaryOp Annotation Name ScalarExpr ScalarExpr

SpecialOp Annotation Name ScalarExprList

AggregateApp Annotation Distinct ScalarExpr ScalarExprDirectionPairList

WindowApp Annotation ScalarExpr ScalarExprList ScalarExprDirectionPairList (Maybe FrameClause)

InPredicate Annotation ScalarExpr Bool InList

LiftApp Annotation Name LiftFlavour ScalarExprList

ScalarSubQuery Annotation QueryExpr

AntiScalarExpr String

Parens Annotation ScalarExpr

data InList

Constructors

InList Annotation ScalarExprList

InQueryExpr Annotation QueryExpr

data LiftFlavour

Constructors

LiftAny

LiftAll

data Direction

Constructors

Asc

Desc

data Distinct

Constructors

Distinct

All

data CombineType

Constructors

Except

Union

Intersect

UnionAll

data IntervalField

Constructors

IntervalYear

IntervalMonth

IntervalDay

IntervalHour

IntervalMinute

IntervalSecond

IntervalYearToMonth

IntervalDayToHour

IntervalDayToMinute

IntervalDayToSecond

IntervalHourToMinute

IntervalHourToSecond

IntervalMinuteToSecond

data ExtractField

Constructors

ExtractCentury

ExtractDay

ExtractDecade

ExtractDow

ExtractDoy

ExtractEpoch

ExtractHour

ExtractIsodow

ExtractIsoyear

ExtractMicroseconds

ExtractMillennium

ExtractMilliseconds

ExtractMinute

ExtractMonth

ExtractQuarter

ExtractSecond

ExtractTimezone

ExtractTimezoneHour

ExtractTimezoneMinute

ExtractWeek

ExtractYear

data FrameClause

Constructors

FrameUnboundedPreceding

FrameUnboundedFull

FrameRowsUnboundedPreceding

Query expressions

data QueryExpr

Constructors

Select

Fields

ann :: Annotation

selDistinct :: Distinct

selSelectList :: SelectList

selTref :: TableRefList

selWhere :: MaybeBoolExpr

selGroupBy :: ScalarExprList

selHaving :: MaybeBoolExpr

selOrderBy :: ScalarExprDirectionPairList

selLimit :: MaybeScalarExpr

selOffset :: MaybeScalarExpr

selOption :: [QueryHint]

CombineQueryExpr

Fields

ann :: Annotation

cqType :: CombineType

cqQe0 :: QueryExpr

cqQe1 :: QueryExpr

Values

Fields

ann :: Annotation

qeValues :: ScalarExprListList

WithQueryExpr

Fields

ann :: Annotation

withs :: WithQueryList

withQe :: QueryExpr

makeSelect :: QueryExpr

'default' valued select, use for creating select values

makeSelect :: QueryExpr
makeSelect = Select
             {ann = emptyAnnotation
             ,selDistinct = All
             ,selSelectList = (SelectList emptyAnnotation [])
             ,selTref = []
             ,selWhere = Nothing
             ,selGroupBy = []
             ,selHaving = Nothing
             ,selOrderBy = []
             ,selLimit = Nothing
             ,selOffset = Nothing
             ,selOption = []}

On its own, it isn't valid syntax: to use it you have to replace the select list at minimum

use something like this

s = makeSelect {selSelectList = sl [se $ i "a"]
               ,selTref = [tref "t"]}
    where
        a = emptyAnnotation
        sl = SelectList a
        se = SelExp a
        i = Identifier a
        tref t = Tref a (Name a [Nmc t])

data WithQuery

Constructors

WithQuery Annotation NameComponent (Maybe [NameComponent]) QueryExpr

data SelectList

Constructors

SelectList Annotation SelectItemList

data SelectItem

Constructors

SelExp Annotation ScalarExpr

SelectItem Annotation ScalarExpr NameComponent

data TableRef

Constructors

Tref Annotation Name

FunTref Annotation ScalarExpr

SubTref Annotation QueryExpr

JoinTref Annotation TableRef Natural JoinType (Maybe JoinHint) TableRef OnExpr

TableAlias Annotation NameComponent TableRef

FullAlias Annotation NameComponent [NameComponent] TableRef

TableRefParens Annotation TableRef

data JoinExpr

Constructors

JoinOn Annotation ScalarExpr

JoinUsing Annotation [NameComponent]

data JoinType

Constructors

Inner

LeftOuter

RightOuter

FullOuter

Cross

data JoinHint

Constructors

Loop

Hash

Merge

data QueryHint

Constructors

QueryHintPartitionGroup

QueryHintColumnarHostGroup

type OnExpr = Maybe JoinExpr

data Natural

Constructors

Natural

Unnatural

Statements

data Statement

Constructors

QueryStatement Annotation QueryExpr

Insert Annotation Name [NameComponent] QueryExpr MaybeSelectList

Update Annotation Name SetClauseList TableRefList MaybeBoolExpr MaybeSelectList

Delete Annotation Name TableRefList MaybeBoolExpr MaybeSelectList

CopyFrom Annotation Name [NameComponent] CopyFromSource [CopyOption]

CopyData Annotation String

CopyTo Annotation CopyToSource String [CopyOption]

Truncate Annotation [Name] RestartIdentity Cascade

CreateTable Annotation Name AttributeDefList ConstraintList MaybeTablePartitionDef

AlterTable Annotation Name AlterTableOperation

AlterDatabase Annotation Name AlterDatabaseOperation

CreateSequence Annotation Name Integer Integer Integer Integer Integer

AlterSequence Annotation Name Name

CreateTableAs Annotation Name QueryExpr

CreateView Annotation Name MaybeNameComponentList QueryExpr

AlterView Annotation Name MaybeNameComponentList QueryExpr

CreateType Annotation Name TypeAttributeDefList

CreateUser Annotation Name String

CreateLogin Annotation Name String

AlterUser Annotation Name String

AlterLogin Annotation Name String

CreateFunction Annotation Name ParamDefList TypeName Replace Language FnBody Volatility

CreateDomain Annotation Name TypeName String MaybeBoolExpr

CreateLanguage Annotation String

CreateTrigger Annotation NameComponent TriggerWhen [TriggerEvent] Name TriggerFire Name ScalarExprList

DropFunction Annotation IfExists NameTypeNameListPairList Cascade

DropSomething Annotation DropType IfExists [Name] Cascade

DropTrigger Annotation IfExists NameComponent Name Cascade

CreateDatabase Annotation Name

Set Annotation String [SetValue]

Notify Annotation String

Into Annotation Bool [Name] Statement

Assignment Annotation Name ScalarExpr

Return Annotation MaybeScalarExpr

ReturnNext Annotation ScalarExpr

ReturnQuery Annotation QueryExpr

Raise Annotation RaiseType String ScalarExprList

NullStatement Annotation

Perform Annotation ScalarExpr

Execute Annotation ScalarExpr

ForQueryStatement Annotation (Maybe String) NameComponent QueryExpr StatementList

ForIntegerStatement Annotation (Maybe String) NameComponent ScalarExpr ScalarExpr StatementList

LoopStatement Annotation (Maybe String) StatementList

WhileStatement Annotation (Maybe String) ScalarExpr StatementList

ContinueStatement Annotation (Maybe String)

ExitStatement Annotation (Maybe String)

CaseStatementSimple Annotation ScalarExpr ScalarExprListStatementListPairList StatementList

CaseStatement Annotation ScalarExprListStatementListPairList StatementList

If Annotation ScalarExprStatementListPairList StatementList

Block Annotation (Maybe String) VarDefList StatementList

AntiStatement String

DeclareStatement Annotation [(String, TypeName, Maybe ScalarExpr)]

ExecStatement Annotation Name ScalarExprList

CreateIndexTSQL Annotation NameComponent Name [NameComponent]

dml components

data CopyToSource

Constructors

CopyTable Name [NameComponent]

CopyQuery QueryExpr

data CopyFromSource

Constructors

CopyFilename String

Stdin

data CopyOption

Constructors

CopyFormat String

CopyDelimiter String

CopyErrorLog String

CopyErrorVerbosity Int

CopyParsers String

data SetClause

Constructors

SetClause Annotation NameComponent ScalarExpr

MultiSetClause Annotation [NameComponent] ScalarExpr

ddl components

data AttributeDef

Constructors

AttributeDef Annotation NameComponent TypeName MaybeScalarExpr RowConstraintList

data RowConstraint

Constructors

NullConstraint Annotation String

NotNullConstraint Annotation String

IdentityConstraint Annotation String (Maybe (Integer, Integer))

RowCheckConstraint Annotation String ScalarExpr

RowUniqueConstraint Annotation String

RowPrimaryKeyConstraint Annotation String

RowReferenceConstraint Annotation String Name (Maybe NameComponent) Cascade Cascade

data Constraint

Constructors

UniqueConstraint Annotation String [NameComponent]

PrimaryKeyConstraint Annotation String [NameComponent]

CheckConstraint Annotation String ScalarExpr

ReferenceConstraint Annotation String [NameComponent] Name [NameComponent] Cascade Cascade

data TablePartitionDef

Constructors

TablePartitionDef Annotation NameComponent Integer TablePartitionDateTimeInterval

data TablePartitionDateTimeInterval

Constructors

Year

Month

Day

Hour

Minute

Second

Millisecond

data TypeAttributeDef

Constructors

TypeAttDef Annotation NameComponent TypeName

data AlterDatabaseOperation

Constructors

RenameDatabase Annotation Name

data AlterTableOperation

Constructors

RenameTable Annotation Name

RenameColumn Annotation NameComponent NameComponent

AlterTableActions Annotation AlterTableActionList

data AlterTableAction

Constructors

AddColumn Annotation AttributeDef

DropColumn Annotation NameComponent

AlterColumn Annotation NameComponent AlterColumnAction

AddConstraint Annotation Constraint

data AlterColumnAction

Constructors

SetDataType Annotation TypeName

SetNotNull Annotation

DropNotNull Annotation

SetDefault Annotation ScalarExpr

DropDefault Annotation

data TriggerWhen

Constructors

TriggerBefore

TriggerAfter

data TriggerEvent

Constructors

TInsert

TUpdate

TDelete

AntiTriggerEvent String

data TriggerFire

Constructors

EachRow

EachStatement

data DropType

Constructors

Table

Domain

View

Type

Database

User

Login

data Cascade

Constructors

Cascade

Restrict

data IfExists

Constructors

Require

IfExists

data RestartIdentity

Constructors

RestartIdentity

ContinueIdentity

function ddl components

data Replace

Constructors

Replace

NoReplace

data Volatility

Constructors

Volatile

Stable

Immutable

data Language

Constructors

Sql

Plpgsql

data FnBody

Constructors

SqlFnBody Annotation StatementList

PlpgsqlFnBody Annotation Statement

PlPgsql components

data ParamDef

Constructors

ParamDef Annotation NameComponent TypeName

ParamDefTp Annotation TypeName

data VarDef

Constructors

VarDef Annotation NameComponent TypeName (Maybe ScalarExpr)

ParamAlias Annotation NameComponent Integer

VarAlias Annotation NameComponent Name

data RaiseType

Constructors

RNotice

RException

RError

utility

data SetValue

Constructors

SetStr Annotation String

SetId Annotation String

SetNum Annotation Double

misc

type WithQueryList = [WithQuery]

type MaybeSelectList = Maybe SelectList

type TableRefList = [TableRef]

type MaybeScalarExpr = Maybe ScalarExpr

type ScalarExprList = [ScalarExpr]

type ScalarExprListList = [ScalarExprTransposedList]

type SetClauseList = [SetClause]

type AttributeDefList = [AttributeDef]

type ConstraintList = [Constraint]

type TypeAttributeDefList = [TypeAttributeDef]

type ParamDefList = [ParamDef]

type TypeNameList = [TypeName]

type NameTypeNameListPair = (Name, TypeNameList)

type NameTypeNameListPairList = [NameTypeNameListPair]

type ScalarExprListStatementListPair = (ScalarExprList, StatementList)

type ScalarExprListStatementListPairList = [ScalarExprListStatementListPair]

type ScalarExprStatementListPair = (ScalarExpr, StatementList)

type ScalarExprStatementListPairList = [ScalarExprStatementListPair]

type VarDefList = [VarDef]

type SelectItemList = [SelectItem]

type RowConstraintList = [RowConstraint]

type CaseScalarExprListScalarExprPair = (ScalarExprList, ScalarExpr)

type CaseScalarExprListScalarExprPairList = [CaseScalarExprListScalarExprPair]

type StatementList = [Statement]

type ScalarExprDirectionPair = (ScalarExpr, Direction)

type ScalarExprDirectionPairList = [ScalarExprDirectionPair]

type AlterTableActionList = [AlterTableAction]

type MaybeNameComponentList = Maybe NameComponentList

type NameComponentList = [NameComponent]

type MaybeBoolExpr = Maybe ScalarExpr